fix: do manual wrapping of body
buildRequestBody checks for an options property and if it's present expects a props to be on a data property. This is not the case with payment requests.

When an options property is present on a payment request the data field in the request becomes undefined. e.g. { data: undefined, options: {...} }

Manual wrapping and then telling the request to not wrap the data prevents this.